CNN Classifier for Dental Implants
In order to replace a lost or broken crown of a dental implant, a dentist has to first virtually identify the type of implant used by the patient from x-ray images.

As there are more than 50 different types of dental implants that are very similar in shape, It is difficult for a dentist to Identify the correct type of implant.

Using CNN and 2363 samples of dental X-rays, the trained model can classify the types of dental implants and detect the correct one with an accuracy of 84%.
The accuracy could reach higher than 90% if we get more trained samples.